Revert "Beta AI transcript reformatter"
This reverts commit 8b8197465a2b1f42e0558b198d6947d556fb3dd9.
